Implemented the feedback
